+++ tests/i2pp/core/test_visualize_results.py
@@ -3,18 +3,13 @@
 from unittest.mock import MagicMock, patch
 
 import pytest
-from i2pp.core.discretization_reader_classes.discretization_reader import (
-    Element,
-)
-from i2pp.core.image_reader_classes.image_reader import (
-    ImageData,
-    PixelValueType,
-)
-from i2pp.core.visualization_classes.discretization_visualization import (
+from i2pp.core.discretization_readers.discretization_reader import Element
+from i2pp.core.image_readers.image_reader import ImageData, PixelValueType
+from i2pp.core.visualize_results import visualize_results, visualize_smoothing
+from i2pp.core.visualizers.discretization_visualizer import (
     DiscretizationVisualizer,
 )
-from i2pp.core.visualization_classes.image_visualization import ImageVisualizer
-from i2pp.core.visualize_results import visualize_results, visualize_smoothing
+from i2pp.core.visualizers.image_visualizer import ImageVisualizer
 
 
 @pytest.fixture
